requirements.txt
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0
httpx>=0.27.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.9.0
